backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
pillow==11.1.0
psycopg2-binary==2.9.10
sqlparse==0.5.3